+++ client/src/components/Common/RequireAuth.tsx
@@ -1,5 +1,5 @@
 import { useLocation, Navigate, Outlet } from "react-router-dom";
-import useAuth from "../hooks/useAuth";
+import useAuth from "../../hooks/useAuth";
 import jwt_decode from "jwt-decode";
 
 interface Props {
