+++ client/src/components/LoginPage/SignIn.tsx
@@ -19,8 +19,8 @@ import {
 } from "@mui/icons-material";
 import { Link as RouterLink, useNavigate, useLocation } from "react-router-dom";
 
-import axios from "../api/axios";
-import useAuth from "../hooks/useAuth";
+import axios from "../../api/axios";
+import useAuth from "../../hooks/useAuth";
 import { useState } from "react";
 
 const LOGIN_URL = "/api/v1/Authentication/Login";
